+++ content/chapters/chapter-1/index.tsx
@@ -18,10 +18,11 @@ export const metadata = {
   title: 'chapter_one.title',
   image: '/assets/images/chapter-1-cover.jpg',
   lessons: [
-    'holocat',
     'genesis-1',
     'genesis-2',
     'genesis-3',
+    'genesis-4',
+    'genesis-5',
     'transacting-1',
     'transacting-2',
     'transacting-3',
