+++ content/lessons/index.tsx
@@ -1,17 +1,19 @@
-import * as Holocat from './holocat'
-import * as Genesis1 from './genesis/genesis-1'
-import * as Genesis2 from './genesis/genesis-2'
-import * as Genesis3 from './genesis/genesis-3'
+import * as Holocat from './genesis/genesis-1'
+import * as Genesis1 from './genesis/genesis-2'
+import * as Genesis2 from './genesis/genesis-3'
+import * as Genesis1Success from './genesis/genesis-4'
+import * as Genesis3 from './genesis/genesis-5'
 import * as Transacting1 from './transacting/transacting-1'
 import * as Transacting2 from './transacting/transacting-2'
 import * as Transacting3 from './transacting/transacting-3'
 import * as Done from './done'
 
 export const lessons = {
-  holocat: Holocat,
-  'genesis-1': Genesis1,
-  'genesis-2': Genesis2,
-  'genesis-3': Genesis3,
+  'genesis-1': Holocat,
+  'genesis-2': Genesis1,
+  'genesis-3': Genesis2,
+  'genesis-4': Genesis1Success,
+  'genesis-5': Genesis3,
   'transacting-1': Transacting1,
   'transacting-2': Transacting2,
   'transacting-3': Transacting3,
